Update API and Views
- Added userCount API endpoint
- Added promote/demote user endpoints and buttons

+++ bot.py
@@ -2,19 +2,20 @@ import asyncio
 import json
 import os
 import time
+from datetime import datetime
 from functools import partial
 from threading import Thread
-from dotenv import load_dotenv
+
 import discord
-import redis
 from bson.json_util import dumps
 from discord.ext import commands
 from dotenv import load_dotenv
 from flask import Flask, jsonify, session, request
 from flask_session import Session
 from requests_oauthlib import OAuth2Session
+
 from utils import ServerSettings, loadAllCogs, loadAllExtensions, UserProfiles, SetupLogger, APIServer, getLogger, \
-    RedisClient, Mongo
+    RedisClient, Mongo, BotAdmins, getSystemLogChannel
 
 # get asyncio loop
 loop = asyncio.get_event_loop()
@@ -92,12 +93,10 @@ async def on_message(message):
         if not server_settings["is_banned"] and not user_profile["MiscData"]["is_banned"]:
             await bot.process_commands(message)
     else:
-        message_responses_enabled = server_settings["modules"]["message_responses"]
-        counting_channel_enabled = server_settings["modules"]["counting_channels"]
 
         if message.channel.type == discord.ChannelType.text and message.channel.name.lower().startswith(
                 "count-to-"):
-            if counting_channel_enabled:
+            if server_settings["modules"]["counting_channels"]:
                 try:
                     # try to convert the string to a number
                     number = int(message.content)
@@ -114,7 +113,7 @@ async def on_message(message):
                     # not a valid number so delete the message
                     await message.delete()
 
-        if message_responses_enabled:
+        if server_settings["modules"]["message_responses"]:
             custom_message_responses = server_settings["custom_message_responses"]
             for custom_response in custom_message_responses:
                 trigger = custom_response["trigger"]
@@ -124,11 +123,31 @@ async def on_message(message):
                     return
 
         # profanity filter
-        words = json.loads(open("settings.json", 'r').read())["profanity_words"]
-        if any(substring in message.content.lower() for substring in words):
-            await message.delete()
-            await asyncio.sleep(1)
-            await message.channel.send("A word you said is not allowed in this server! :rage:")
+        if server_settings["modules"]["profanity_filter"]:
+            words = json.loads(open("settings.json", 'r').read())["profanity_words"]
+            if any(substring in message.content.lower() for substring in words):
+                await message.delete()
+                await asyncio.sleep(1)
+                message = await message.channel.send("A word you said is not allowed in this server! :rage:")
+                await asyncio.sleep(5)
+                await message.delete()
+
+
+@app.route("/api/v1/userCount", methods=["GET"])
+def api_user_count():
+    if bot.is_ready():
+        token = request.headers.get("Token")
+        if token is not None:
+            token = json.loads(token)
+            discord_session = make_session(token=token)
+            if discord_session.authorized:
+                return jsonify({"user_count": len(bot.users)}), 200
+            else:
+                return "", 401
+        else:
+            return "", 403
+    else:
+        return "bot is not ready!", 500
 
 
 @app.route("/api/v1/users", methods=["GET"])
@@ -146,7 +165,8 @@ def api_users():
                     x = bot.get_user(user_json["id"])
                     users.append({"username": x.name, "id": x.id, "discriminator": x.discriminator,
                                   "avatar_url": str(x.avatar_url),
-                                  "is_banned": user_json["MiscData"]["is_banned"]})
+                                  "is_banned": user_json["MiscData"]["is_banned"],
+                                  "is_admin": user["id"] in BotAdmins().get()})
                 return jsonify(users)
             else:
                 return "", 401
@@ -243,6 +263,13 @@ def admin_ban_server_notify():
                                 loop
                             )
                             send_fut.result()
+
+                            send_fut2 = asyncio.run_coroutine_threadsafe(
+                                getSystemLogChannel(bot).send(
+                                    f"{guild.name} ({guild.id}) was banned by ``{banner_user['username']}``."),
+                                loop
+                            )
+                            send_fut2.result()
                             return "OK", 200
                         except discord.Forbidden as e:
                             getLogger().error(
@@ -288,6 +315,13 @@ def admin_unban_server_notify():
                                 loop
                             )
                             send_fut.result()
+
+                            send_fut2 = asyncio.run_coroutine_threadsafe(
+                                getSystemLogChannel(bot).send(
+                                    f"{guild.name} ({guild.id}) was unbanned by ``{banner_user['username']}``."),
+                                loop
+                            )
+                            send_fut2.result()
                             return "OK", 200
                         except discord.Forbidden as e:
                             getLogger().error(
@@ -334,6 +368,13 @@ def admin_ban_user_notify():
                                 loop
                             )
                             send_fut.result()
+
+                            send_fut2 = asyncio.run_coroutine_threadsafe(
+                                getSystemLogChannel(bot).send(
+                                    f"[{user.mention}] {user} ({user.id}) was banned by ``{banner_user['username']}``."),
+                                loop
+                            )
+                            send_fut2.result()
                             return "OK", 200
                         except discord.Forbidden as e:
                             getLogger().error(
@@ -379,6 +420,13 @@ def admin_unban_user_notify():
                                 loop
                             )
                             send_fut.result()
+
+                            send_fut2 = asyncio.run_coroutine_threadsafe(
+                                getSystemLogChannel(bot).send(
+                                    f"[{user.mention}] {user} ({user.id}) was unbanned by ``{banner_user['username']}``."),
+                                loop
+                            )
+                            send_fut2.result()
                             return "OK", 200
                         except discord.Forbidden as e:
                             getLogger().error(
@@ -417,7 +465,7 @@ def admin_toggle_module():
                     enabled = request.get_json()["enabled"]
 
                     old_json = json.loads(open("settings.json", 'r').read())
-                    old_json["modules"][module] = enabled
+                    old_json["modules"][module]["enabled"] = enabled
 
                     with open("settings.json", 'w') as f:
                         f.write(json.dumps(old_json))
@@ -441,7 +489,7 @@ def admin_modules():
             token = json.loads(token)
             discord_session = make_session(token=token)
             if discord_session.authorized:
-                return jsonify(json.loads(open("settings.json", 'r').read()))
+                return jsonify(json.loads(open("settings.json", 'r').read())["modules"])
             else:
                 return "", 401
         else:
@@ -459,6 +507,7 @@ def api_get_server(server_id):
             discord_session = make_session(token=token)
             if discord_session.authorized:
                 server = bot.get_guild(server_id)
+                server_settings = ServerSettings(server).getServerDocument()
                 if server is not None:
                     return jsonify(
                         {"name": server.name, "id": server.id, "region": server.region.name,
@@ -468,7 +517,8 @@ def api_get_server(server_id):
                          "category_amount": len(server.categories), "member_count": server.member_count,
                          "role_amount": len(server.roles),
                          "text_channels": [{"name": y.name, "id": y.id} for y in server.text_channels],
-                         "roles": [{"name": z.name, "id": z.id} for z in server.roles if z.name != "@everyone"]})
+                         "roles": [{"name": z.name, "id": z.id} for z in server.roles if z.name != "@everyone"],
+                         "log_channel": server_settings["log_channel"], "is_banned": server_settings["is_banned"]})
                 else:
                     return "", 400
             else:
@@ -522,6 +572,138 @@ def api_get_user(user_id):
             return "", 403
     else:
         return "bot is not ready!", 500
+
+
+@app.route("/api/v1/admin/promoteUser", methods=["POST"])
+def admin_promote_user():
+    if request.is_json:
+        if bot.is_ready():
+            token = request.headers.get("Token")
+            if token is not None:
+                token = json.loads(token)
+                discord_session = make_session(token=token)
+                if discord_session.authorized:
+                    user_id = int(request.get_json()["user_id"])
+                    user = bot.get_user(user_id)
+                    if user is not None:
+                        BotAdmins().add(user_id)
+                        try:
+                            promoter = discord_session.get("https://discordapp.com/api/users/@me").json()
+                            promoter_user = bot.get_user(int(promoter['id']))
+
+                            user_embed = discord.Embed(title="You have been promoted to an Admin of mine!",
+                                                       description=f"You will now have access to the admin panel of the dashboard and bot admin commands!\nThis promotion was issued by ``{promoter['username']}``.",
+                                                       color=discord.Color.green(),
+                                                       timestamp=datetime.utcnow())
+                            user_embed.set_author(name=bot.user.name, icon_url=bot.user.avatar_url)
+
+                            send_fut = asyncio.run_coroutine_threadsafe(
+                                user.send(content=None, embed=user_embed),
+                                loop
+                            )
+                            send_fut.result()
+
+                            log_embed = discord.Embed(title="New Bot Admin",
+                                                      description=f"{user.mention} was added as a bot admin by {promoter_user.mention}",
+                                                      color=discord.Color.green(),
+                                                      timestamp=datetime.utcnow())
+                            log_embed.set_author(name=bot.user.name, icon_url=bot.user.avatar_url)
+                            # embed.set_footer(text=author.name, icon_url=author.avatar_url)
+
+                            send_fut2 = asyncio.run_coroutine_threadsafe(
+                                getSystemLogChannel(bot).send(content=None, embed=log_embed),
+                                loop
+                            )
+                            send_fut2.result()
+                            return "OK", 200
+                        except discord.Forbidden as e:
+                            getLogger().error(
+                                f"Missing permission to send promotion notification to user {user} ({user.id})! Error: {e.text}")
+                            return "missing permissions", 500
+                        except discord.HTTPException as e:
+                            getLogger().error(
+                                f"Failed to send promotion notification to user {user} ({user.id})! Error: {e.text}")
+                            return "failed", 500
+                        except discord.InvalidArgument as e:
+                            getLogger().error(
+                                f"Failed to send promotion notification to user {user} ({user.id})! Error: {e}")
+                            return "invalid argument, we shouldnt be here?!", 500
+                    else:
+                        return "user is none", 400
+                else:
+                    return "", 401
+            else:
+                return "", 403
+        else:
+            return "bot is not ready!", 500
+    else:
+        return "", 400
+
+
+@app.route("/api/v1/admin/demoteUser", methods=["POST"])
+def admin_demote_user():
+    if request.is_json:
+        if bot.is_ready():
+            token = request.headers.get("Token")
+            if token is not None:
+                token = json.loads(token)
+                discord_session = make_session(token=token)
+                if discord_session.authorized:
+                    user_id = int(request.get_json()["user_id"])
+                    user = bot.get_user(user_id)
+                    if user is not None:
+                        BotAdmins().remove(user_id)
+                        try:
+                            demoter = discord_session.get("https://discordapp.com/api/users/@me").json()
+                            demoter_user = bot.get_user(int(demoter['id']))
+
+                            user_embed = discord.Embed(title="You have been demoted from an Admin of mine!",
+                                                       description=f"You will no longer have access to the admin panel of the dashboard and bot admin commands!\nThis demotion was issued by ``{demoter['username']}``.",
+                                                       color=discord.Color.red(),
+                                                       timestamp=datetime.utcnow())
+                            user_embed.set_author(name=bot.user.name, icon_url=bot.user.avatar_url)
+
+                            send_fut = asyncio.run_coroutine_threadsafe(
+                                user.send(content=None, embed=user_embed),
+                                loop
+                            )
+                            send_fut.result()
+
+                            log_embed = discord.Embed(title="Bot Admin Removed",
+                                                      description=f"{user.mention} was removed as a bot admin by {demoter_user.mention}",
+                                                      color=discord.Color.red(),
+                                                      timestamp=datetime.utcnow())
+                            log_embed.set_author(name=bot.user.name, icon_url=bot.user.avatar_url)
+                            # embed.set_footer(text=author.name, icon_url=author.avatar_url)
+
+                            send_fut2 = asyncio.run_coroutine_threadsafe(
+                                getSystemLogChannel(bot).send(content=None, embed=log_embed),
+                                loop
+                            )
+                            send_fut2.result()
+                            return "OK", 200
+                        except discord.Forbidden as e:
+                            getLogger().error(
+                                f"Missing permission to send demotion notification to user {user} ({user.id})! Error: {e.text}")
+                            return "missing permissions", 500
+                        except discord.HTTPException as e:
+                            getLogger().error(
+                                f"Failed to send demotion notification to user {user} ({user.id})! Error: {e.text}")
+                            return "failed", 500
+                        except discord.InvalidArgument as e:
+                            getLogger().error(
+                                f"Failed to send demotion notification to user {user} ({user.id})! Error: {e}")
+                            return "invalid argument, we shouldnt be here?!", 500
+                    else:
+                        return "user is none", 400
+                else:
+                    return "", 401
+            else:
+                return "", 403
+        else:
+            return "bot is not ready!", 500
+    else:
+        return "", 400
 
 
 def token_updater(token):
